Updated the logic
Updated the logic for date.

On selecting End date as today's date, the cumulative data appears. But on selecting, End date as the last day of the month, the data is appearing for that specific month.
Earlier, in practice it was like, on selecting today's date as "End date", only the data used to appear for the month which is selected in the "Start date".